Communication betwee Net Calculator and Tax Rate Provider established.

+++ net-calculator/src/main/java/gsg/nc/calc/service/NetCalculatorService.java
@@ -3,5 +3,5 @@ package gsg.nc.calc.service;
 import java.math.BigDecimal;
 
 public interface NetCalculatorService {
-    public BigDecimal calculateNetPrice(BigDecimal grossPrice, String countryIso);
+    public double calculateNetPrice(double grossPrice, String countryIso);
 }
